% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{code}
\alias{code}
\title{Construct coded name}
\usage{
code(strs)
}
\arguments{
\item{strs}{character vector}
}
\value{
coded variable name
}
\description{
Construct coded name
}
\examples{
code(c("GF", "A20to24"))

}

