% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct.R
\name{construct}
\alias{construct}
\title{Construct new measures.}
\usage{
construct(name, data)
}
\arguments{
\item{name}{construction name.}

\item{data}{ACS dataset.}
}
\value{
A new construction.
}
\description{
Construct new measures.
}
\examples{
construct("Child Poverty", data = acs_data)

}

