% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_acs_data.R
\name{fetch_acs_data}
\alias{fetch_acs_data}
\title{Fetch ACS dataset.}
\usage{
fetch_acs_data(name, geo, endyear, span)
}
\arguments{
\item{name}{construction name.}

\item{geo}{geography definition.}

\item{endyear}{survey end year.}

\item{span}{survey span.}
}
\value{
An ACS dataset.
}
\description{
Fetch ACS dataset.
}

