% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{code}
\alias{code}
\title{Construct coded name}
\usage{
code(strs, year)
}
\arguments{
\item{strs}{character vector}

\item{year}{survey year}
}
\value{
coded variable name
}
\description{
Construct coded name
}
\examples{
code(c("Amrs", "LatAm", "Crb", "Wid3"), 2014)

}

